# Conflicts:
#	Android/core/data/src/main/java/com/ssafy/data/repository/board/BoardRepositoryImpl.kt
#	Android/core/model/src/main/java/com/ssafy/model/board/BoardDTO.kt

+++ Android/feature/home/src/main/java/com/ssafy/home/home/HomeDestination.kt
@@ -12,7 +12,7 @@ fun NavGraphBuilder.homeScreen(
     moveToBoardScreen: (Long) -> Unit,
     moveToCreateNewBoardScreen: (List<WorkSpaceDTO>) -> Unit,
     moveToLoginScreen: () -> Unit,
-    moveToSettingScreen: () -> Unit,
+    moveToSettingScreen: (Long) -> Unit,
     moveToMyCardScreen: () -> Unit,
     moveToUpdateProfile: () -> Unit,
     moveToSearchScreen: () -> Unit,
